+++ first-task/src/pages/Main/Main.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Card } from 'components/Card/Card';
 import { Header } from 'components/Header/Header';
-import { SateCharacters, Character } from '../typingData';
+import { SateCharacters, Character } from '../../components/typingData';
 import styles from './Main.module.css';
 
 const BASE_PATH = 'https://rickandmortyapi.com/api/character';
